Added rudimentary support for thread-local storage.
We don't actually use that in M³ and also don't plan to, but it is
required for exception handling in C++, because libsupc++ makes use of
TLS when not build with --disable-tls. We used to specify that flag for
our own cross compiler, but now with buildroot we want to build it just
like Linux needs it.