riscv32: make time_t 32-bit.
This is actually a workaround for a problem in Rust's libc crate, which
defines time_t as a 32-bit type. musl changed this type for all 32-bit
architectures to 64-bit in 1.2 apparently and the libc wasn't adapted
yet. There is already an issue:

https://github.com/rust-lang/libc/issues/1848

Until this is resolved, we change the type in musl to 32-bit to avoid
the issues when using the Rust standard library.