Added workaround to clock_gettime for rv32 change.
Now that we define time_t to be 32-bit on riscv32, we ran into an issue
in clock_gettime which had a special handling for the case that time_t
is larger than 32-bit. In the end, it made the clock_gettime syscall
fail.

I've therefore added a workaround to repair that for us. Since it does
not matter which clock_gettime (normal or 64-version) we use, we can
just always call clock_gettime64.